Project Downgraded to Unity 2021 LTS
The project was deep in development when I noticed I was using the wrong version of Unity. I downgraded it from Unity 2022 LTS to Unity 2021.
I had to remove the com.unity.feature.2d: Package [com.unity.feature.2d@2.0.0] and restart the project to fix the project. I also had to adjust the bounds of the Box Collider components that are inside of the card prefabs.